% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRoadFromPath.R
\name{buildRoadFromPath}
\alias{buildRoadFromPath}
\title{build road based on paths}
\usage{
buildRoadFromPath(sim)
}
\arguments{
\item{sim}{a sim list

Changed from roadCLUS.analysis}
}
\description{
build road based on paths
}
