% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSnapRoads.R
\name{buildSnapRoads}
\alias{buildSnapRoads}
\title{Build roads with the snap method}
\usage{
buildSnapRoads(sim)
}
\arguments{
\item{sim}{}
}
\value{
sim list
}
\description{
Build roads that simply connect each landing to each road "as the crow
flies."
}
