% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLUSexample.R
\docType{data}
\name{CLUSexample}
\alias{CLUSexample}
\title{Data from the CLUS example}
\format{A named list where: $cost is an object of class RasterLayer representing road cost
                            $landings is an oject of class SpatialPoints representing landing locations
                            $roads is an object of class RasterLayer representing existing roads}
\usage{
data(CLUSexample)
}
\description{
Data from the CLUS example
}
\keyword{datasets}
