% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraph.R
\name{getGraph}
\alias{getGraph}
\title{Set the graph which determines least cost paths}
\usage{
getGraph(sim, neighbourhood)
}
\arguments{
\item{sim}{a sim list}

\item{neighbourhood}{neighbourhood type}
}
\description{
Creates a graph in initiation phase which can be updated and solved for
paths
}
