% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLandingsFromTarget.R
\name{getLandingsFromTargetRast}
\alias{getLandingsFromTargetRast}
\title{Select random landing locations within patches.}
\usage{
getLandingsFromTargetRast(
  inputPatches,
  landingDens,
  sampleType = "regular",
  omitCentroidsOutOfPolygons = F
)
}
\arguments{
\item{inputPatches}{A RasterLayer. Harvested patches should have values
greater than 0}

\item{landingDens}{number of landings per unit area. This should be in the
same units as the crs of the harvest. Note that 0.001 pts per m2 is > 1000
pts per km2 so this number is usually very small for projected crs.}

\item{sampleType}{character. "centroid" (default), "regular" or "random".
Centroid returns one landing per harvest block, which is not guaranteed to
be in the harvest block. Regular returns points from a grid with density
\code{landingDens} that overlap the harvested areas. Random returns a
random set of points from each polygon where the number is determined by
the area of the polygons and \code{landingDens}. The centroid is always
returned as one of the landings to ensure all harvest areas get at least
one landing.}

\item{omitCentroidsOutOfPolygons}{Logical. Default is FALSE in which case
some points may be outside the borders of the harvested patch}
}
\description{
Select random landing locations within patches.
}
