% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathsToLines.R
\name{pathsToLines}
\alias{pathsToLines}
\title{Convert raster cell numbers in paths to lines}
\usage{
pathsToLines(sim)
}
\arguments{
\item{sim}{sim list}
}
\description{
copied from newRoadsToLines and made work with sf
}
