% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepExData.R
\name{prepExData}
\alias{prepExData}
\title{Prepare example data}
\usage{
prepExData(x)
}
\arguments{
\item{x}{list. Contains elements some of which are packed \code{SpatRasters}.}
}
\value{
The same list but with unwrapped \code{SpatRasters}
}
\description{
Prepare example data included in the package that contain wrapped terra
objects. This applies \code{\link[terra:wrap]{terra::unwrap()}} recursively to the list provided so
that all \code{PackedSpatRasters} are converted to \code{SpatRasters}.
}
\examples{
CLUSexample
prepExData(CLUSexample)
}
