% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectRoadsNew.R
\name{projectRoadsNew}
\alias{projectRoadsNew}
\title{Project road network}
\usage{
projectRoadsNew(
  landings = NULL,
  cost = NULL,
  roads = NULL,
  roadMethod = "mst",
  plotRoads = F,
  mainTitle = NULL,
  neighbourhood = "octagon",
  sim = NULL
)
}
\arguments{
\item{landings}{}

\item{cost}{}

\item{roads}{}

\item{roadMethod}{}

\item{plotRoads}{}

\item{mainTitle}{}

\item{neighbourhood}{}

\item{sim}{}
}
\description{
Project road locations based on existing roads, planned landings, and a cost
surface that defines the cost of building roads.
}
\details{
Three different methods for projecting road networks have been implemented:
 \itemize{
 \item{"snap":} {connects each landing directly to the closest road without
  reference to the cost or other landings}
 \item{"lcp":} {Least Cost Path connects each landing to the closest point on 
 the road by determining the least cost path based on the cost surface 
 provided, it does not consider other landings}
 \item{"mst":} {Minimum Spanning Tree connects all landings to the road by 
 determining the least cost path to the road or other landings based on the 
 cost surface}
 }
}
