% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterToLineSegments.R
\name{rasterToLineSegments}
\alias{rasterToLineSegments}
\title{Convert raster to lines}
\usage{
rasterToLineSegments(rast)
}
\arguments{
\item{rast}{raster with lines}
}
\description{
Doesn't work perfectly
}
