% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortestPaths.R
\name{shortestPaths}
\alias{shortestPaths}
\title{Get shortest path between points in list}
\usage{
shortestPaths(sim)
}
\arguments{
\item{sim}{sim list}
}
\description{
Get shortest path between points in list
}
