% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFnsClus.R
\name{simpleCost}
\alias{simpleCost}
\title{Creat a simple cost layer}
\usage{
simpleCost(roads, newLandings, water)
}
\arguments{
\item{roads}{roads lines or raster}

\item{newLandings}{landings raster}

\item{water}{water raster}
}
\description{
Creat a simple cost layer
}
