% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightFunctions.R
\name{simpleCostFn}
\alias{simpleCostFn}
\title{Simple cost edge weight function}
\usage{
simpleCostFn(x1, x2, hdistance)
}
\arguments{
\item{x1, x2}{Number. Value of the input cost raster at two nodes.}

\item{hdistance}{Number. Horizontal distance between the nodes - for penalizing longer diagonal edges.}
}
\description{
Calculates the weight of an edge between two nodes as the mean value
of an input cost raster at each of those nodes (x1 and x2).
}
\examples{
simpleCostFn(0.5,0.7,1)
}
