% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSimList.R
\name{updateSimList}
\alias{updateSimList}
\title{Update and existing sim list}
\usage{
updateSimList(sim, landings)
}
\arguments{
\item{sim}{sim list from input}

\item{landings}{input landigs}
}
\description{
Update sim list with new landings and burn in roads to have 0 cost
}
