% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSimList.R
\name{buildSimList}
\alias{buildSimList}
\title{Build sim list object}
\usage{
buildSimList(roads, cost, roadMethod, landings)
}
\description{
Build a sim list object to be used in the rest of the functions and returned
by \code{projectRoads}. It will convert other types of spatial objects to sf
or raster.
}
