% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLUSexample.R
\docType{data}
\name{CLUSexample}
\alias{CLUSexample}
\title{Data from the CLUS example}
\format{
A named list with components:
\itemize{
\item cost: a \code{PackedSpatRaster} representing road building cost.
\item landings: an sf dataframe of points representing landing locations.
\item roads: a \code{PackedSpatRaster} representing existing roads.
}
}
\usage{
data(CLUSexample)
}
\description{
From Kyle Lochhead and Tyler Muhly's CLUS road simulation example. \code{SpatRaster}
files created with the terra package must be saved with \code{\link[terra:wrap]{terra::wrap()}} and
need to be unwrapped before they are used. \code{\link[=prepExData]{prepExData()}} does this.
}
\examples{
CLUSexample
prepExData(CLUSexample)
}
\keyword{datasets}
