% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFnsClus.R
\name{getCentroids}
\alias{getCentroids}
\title{Get centroids from raster landings}
\usage{
getCentroids(newLandings, withIDs = T)
}
\arguments{
\item{newLandings}{raster landings}

\item{withIDs}{logical}
}
\description{
Get centroids from raster landings
}
