% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClosestRoad.R
\name{getClosestRoad}
\alias{getClosestRoad}
\title{Find the closest point on a road for each landing}
\usage{
getClosestRoad(sim)
}
\arguments{
\item{sim}{sim list}
}
\description{
Find the closest point on the road to each landing and return as an x y
matrix
}
