% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistFromSource.R
\name{getDistFromSource}
\alias{getDistFromSource}
\title{Moving window approach to get distance from roads}
\usage{
getDistFromSource(src, maxDist, kwidth = 3, dissag = F)
}
\arguments{
\item{dissag}{}
}
\value{

}
\description{
There is a tradeoff between speed and accuracy in this method, controlled by
the parameters kwidth and dissag. kwidth is the radius of the moving window
in number of cells. Smaller values will provide a smoother gradient while
larger values will give a blockier result. If dissag is false, the resulting
distance map will have values of 0, 1kwidth, 2kwidth, etc. If dissag is true
the algorithm disaggregates the raster to a finer resolution, so the
resulting distance map will have values of 0,1,2, etc. Setting dissag = TRUE
or a low value for kwidth will increase the computation time.
}
