% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGraph.R
\name{getGraph}
\alias{getGraph}
\title{Set the graph which determines least cost paths}
\usage{
getGraph(sim, neighbourhood)
}
\description{
Creates a graph in inititation phase which can be updated and solved for paths
}
