% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLandingsFromTarget.R
\name{getLandingsFromHarvest}
\alias{getLandingsFromHarvest}
\title{Get landing points inside harvest blocks}
\usage{
getLandingsFromHarvest(harvest, numLandings = 1)
}
\description{
Get landing points inside harvest blocks
}
