% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLandingsFromTarget.R
\name{getLandingsFromTarget}
\alias{getLandingsFromTarget}
\title{Select random landing locations within patches.}
\usage{
getLandingsFromTarget(
  inputPatches,
  numLandings,
  omitCentroidsOutOfPolygons = F
)
}
\arguments{
\item{inputPatches}{A RasterLayer. Harvested patches should have values equal
to a unique identifier}

\item{numLandings}{Numeric. A vector of the number of points to randomly
sample inside harvested patches}

\item{omitCentroidsOutOfPolygons}{Logical. Default is FALSE in which case
some points may be outside the borders of the harvested patch}
}
\description{
Select random landing locations within patches.
}
