% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepExData.R
\name{prepExData}
\alias{prepExData}
\title{Prepare example data}
\usage{
prepExData(x)
}
\arguments{
\item{x}{A list containing elements some of which are packed SpatRasters.}
}
\value{
The same list but with unwrapped SpatRasters
}
\description{
Prepare example data included in the package that contain wrapped terra
objects. This applies [terra::unwrap()] recursively to the list provided so
that all PackedSpatRasters are converted to SpatRasters.
}
\examples{
CLUSexample
prepExData(CLUSexample)
}
