% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectRoads.R
\docType{methods}
\name{projectRoads}
\alias{projectRoads}
\alias{projectRoads,matrix-method}
\alias{projectRoads,RasterLayer-method}
\alias{projectRoads,SpatialPolygons-method}
\alias{projectRoads,RasterStack-method}
\alias{projectRoads,RasterBrick-method}
\title{Project road network.}
\usage{
projectRoads(landings, cost = NULL, roads = NULL, roadMethod = "mst",
  plotRoads = T, sim = list())

\S4method{projectRoads}{matrix}(landings, cost = NULL, roads = NULL,
  roadMethod = "mst", plotRoads = T, sim = list())

\S4method{projectRoads}{RasterLayer}(landings, cost = NULL,
  roads = NULL, roadMethod = "mst", plotRoads = T, sim = list())

\S4method{projectRoads}{SpatialPolygons}(landings, cost = NULL,
  roads = NULL, roadMethod = "mst", plotRoads = T, sim = list())

\S4method{projectRoads}{RasterStack}(landings, cost = NULL,
  roads = NULL, roadMethod = "mst", plotRoads = T, sim = list())

\S4method{projectRoads}{RasterBrick}(landings, cost = NULL,
  roads = NULL, roadMethod = "mst", plotRoads = T, sim = list())
}
\arguments{
\item{landings}{RasterLayer, SpatialPolygons*, SpatialPointsDataFrame*, matrix, or RasterStack. Features to be connected to the road network. matrix should contain x,y,v columns, as returned by rasterToPoints etc. If RasterStack assume an ordered time-series.}

\item{cost}{RasterLayer. Cost surface.}

\item{roads}{RasterLayer. Existing road network.}

\item{roadMethod}{Character. Options are "mst", "lcp", "snap".}

\item{plotRoads}{Boolean. Set FALSE to save time if output road rasters are not required. Default TRUE}

\item{sim}{Sim list. Returned from a previous iteration of projectRoads. cost, roads, and roadMethod are ignored if a sim list is provided.}
}
\value{
sim list.

sim list.

sim list.

RasterBrick. Road network over time.

RasterBrick. Road network over time.
}
\description{
Project road network.
}
\details{

}
\examples{
# TODO: examples
}
